+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.3
+- Added for Android: Print over Bluetooth® Low Energy
+
 ## 0.0.2+5
 
 - Fix duplicate on Android